refactor: inline SessionResources type and remove backend-ai-console.d.ts
Move SessionResources interface into backend.ai-client-esm.ts where it is
solely used, eliminating the now-unnecessary type declaration file.
