--- src/types/backend-ai-console.d.ts
@@ -1,38 +1,0 @@
-// The BackendAIWebUI Lit shell has been removed.
-// Login flow, logout, global stores, and language sync are all handled by React.
-
-export interface SessionResources {
-  group_name?: string;
-  domain?: string;
-  type?: 'interactive' | 'batch' | 'inference' | 'system';
-  cluster_mode: 'single-node' | 'multi-node';
-  cluster_size: number;
-  maxWaitSeconds: number;
-  starts_at?: string;
-  startupCommand?: string;
-  bootstrap_script?: string;
-  owner_access_key?: string;
-  reuseIfExists?: boolean;
-  config?: {
-    resources?: {
-      cpu: number;
-      mem: string;
-      [key: string]: number | string;
-    };
-    resource_opts?: {
-      shmem?: string;
-      allow_fractional_resource_fragmentation?: boolean;
-    };
-    mounts?: string[];
-    mount_ids?: string[];
-    mount_map?: {
-      [key: string]: string;
-    };
-    environ?: {
-      [key: string]: string;
-    };
-    scaling_group?: string;
-    preopen_ports?: number[];
-    agent_list?: string[];
-  };
-}
